# rust/api/openapi.yml
openapi: "3.1.1"
info:
  title: "Scanner"
  description: "
    # Authentication

    The API supports two kinds of authentication methods:

    - API Key

    - Certificates


    The authentication modes are set within a configuration file or via the argument list, when starting the server.

    The authentication is required for each request except for a HEAD request.

    ## API Key

    An API key is a token that the client provides when doing API requests and are used to authorize access.
    The `X-API-KEY` must be in the header.


    <!--More details about this method follows with its implementation.-->


    ## Certificates

    This option uses [X.509](https://en.wikipedia.org/wiki/X.509), based on CA to verify derived certificates to allow access.


    <!--More details about this method follows with its implementation.-->
    "
  contact:
    name: "Greenbone AG"
    url: "https://www.greenbone.net/"
  license:
    name: "GPL-2.0-or-later"
    identifier: "GPL-2.0-or-later"
    url: "https://spdx.org/licenses/GPL-2.0-or-later.html"
  version: "0.1"
servers:
  - url: /
...
paths:
...
  /vts:
    head:
      description: "Get the response header. It contains the API version, feed version and available authentication methods."
      operationId: "head_vts"
      tags:
        - "general"
      responses:
        "200":
          headers:
            api-version:
              $ref: "#/components/headers/ApiVersion"
            feed-version:
              $ref: "#/components/headers/FeedVersion"
            authentication:
              $ref: "#/components/headers/Authentication"
          description: "Authenticated and authorized"
    get:
      description:
        "Get a list of all available vulnerability tests (VTs) to the scanner. Note, that not the
        tests itself, but a list of identifiers (OIDs) of the tests is returned."
      operationId: "get_vts"
      tags:
        - "feed"
      responses:
        "200":
          description: "A list of available VTs."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: "string"
              examples:
                list of OIDs:
                  $ref: "./components/examples/vts.yml#/list_of_oids"
        "503":
          description: "The list of OIDs is currently updated. Please try again later."
components:
  headers:
    ApiVersion:
      $ref: "./components/headers.yml#/ApiVersion"
    FeedVersion:
      $ref: "./components/headers.yml#/FeedVersion"
    Authentication:
      $ref: "./components/headers.yml#/Authentication"
...